Add a confd client function for name of diagnose collector
The monitoring daemon needs to be aware of the name of the node
status collector in order to provide accurate node status data.
The way for the monitoring daemon to do this, is to query the
configuration daemon; this is necessary, as the monitoring daemon
runs on every node, so cannot assume to have a local copy of the
configuration.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Viktor Bachraty <vbachraty@google.com>