diagnose status collector: take name from confd
The diagnose status collector gets the name of the diagnose
program from the configuration. (Of course, independent
white-listing by the node is checked afterwards.) The standard
way for MonD to access the configuration is by asking the ConfDs;
this is necessary, as MonD is present on any node, but the
configuration only on the master node. Do this.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Viktor Bachraty <vbachraty@google.com>