Optimize LXC hypervisor GetAllInstancesInfo
Don't list all running instances for each instance, generate the list
only once at start. Unfortunately, listing (as most other operations)
is done by executing a cmdline tool in bash, which is too slow for
frequent operations. This caused RPC to time out after 60s in case there
are >200 instances, with this patch it is <1s.

Signed-off-by: Viktor Bachraty <vbachraty@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>