Improve luxid QueryInstances performance for large clusters
Do this by implementing and using getInstanceByName on hot paths.

This is much faster than the more generic but extremely inefficient
getInstance which can match by UUID, exact name, or name prefixes.
instances are stored in a Map keyed by UUID, and if that lookup fails,
getInstance creates a temporary Map of names to instances, and
does a linear scan of it.

On a large test cluster (80 nodes, 1000 instances), getInstance used
~65% of the total CPU time and 80% of the heap allocations for a luxid
that started, handled 2 QueryInstance calls, and shut down.

This patch reduces the time to respond to the client for each of these
calls from ~18s to ~2.5s, and the total luxid heap allocations from
~11.5GB to ~2.4GB.

Signed-off-by: Brian Foley <bpfoley@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>