Fix hlint errors
Prefer using pointfree style, remove unnecessary parentheses. There
is a bug in hlint (https://github.com/ndmitchell/hlint/issues/196) that
turns the "role" indentifier into a keyword. It can be disabled using
-XNoRoleAnnotations but this flag is not supported by ghc 7.4.x which
is distributed with wheezy. It could be worked around by detecting ghc
version in automake.am, but a much simpler solution is to avoid using
this identifier.

Signed-off-by: Viktor Bachraty <vbachraty@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>