Reduce load in NV_NODENETTEST and NV_MASTERIP
Run these tests by master candidates only. In larger clusters, these
test were effectively making a DoS attack on the master node.
LUClusterVerify() calls node_verify on each node, that would run
TCPPing on noded's port for everyone in the nodegroup + master ip. This
was causing timeouts on the master node being unable to receive
node_verify RPC. Reducing this test to master candidates can reduce the
number of connections to noded on the master node by a factor of 8.

Signed-off-by: Viktor Bachraty <vbachraty@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>