Cancel RAPI job if the client drops the connection
This commit fixes a bug where long-lasting OsDiagnose jobs would
unnecessarily execute after a RAPI client had already closed the
connection. This fix introduces a couple new polling/timeout
constants for Luxi and RAPI to handle the polling of the HTTP
connection while at the same time monitor the state of the submitted
job.

If the connection is closed, client-side, the job is canceled and it
will be removed from the job queue if not already in a running state.

Signed-off-by: Federico Morg Pareschi <morg@google.com>
Signed-off-by: Viktor Bachraty <vbachraty@google.com>
Reviewed-by: Viktor Bachraty <vbachraty@google.com>