Reduce heap when parsing & storing ConfigData 10%
Do this by changing the string representation for field names and field
documentation in the Field data type from String to Data.Text. This type
is used primarily in the Template Haskell JSValue -> data type code.

This representation is only used internally in two places:
1. In template haskell helper functions used to generate splices.
2. The UsedKeys type and fieldDictKeys is used in readJSONfromDict
   to verify that a JSON object doesn't contain any unexpected
   ('superfluous') keys.

Building this UsedKey list set for each data type mapped onto JSON uses
a substantial amount of heap and occurs in luxi, confd, and wconfd.
For a simple harness that parses a large config.data into ConfigData,
then GCs, the live heap reported by +RTS -S after a manual full GC is
46.9MB, the majority of which is in String values.

If we completely disable the 'superfluous' checking, this reduces to
38.5MB, or a saving of 8.4MB, because fieldDictKeys is never evaluated.

With 'superfluous' checking enabled, but Data.Text fields, this becomes
42.6MB, or a saving of 4.3MB.

Signed-off-by: Brian Foley <bpfoley@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>