Use socat method string compatible with <1.73 & >=1.73
As reported in issue 1172, socat 1.7.3 changed the list of protocol
version names for TLS 1.0 from "TLS", "TLS1", "TLSv1" to ["TLS1" and
 "TLS1.0"].

socat 1.7.3 was released in January 2015 and is used in Debian Jessie
and later; and Ubuntu Wily (15.10) and later.

https://github.com/craSH/socat/commit/d430147bc60c

Update Ganeti to use "TLS1" which is compatible with both, and works
all the way back to at least socat 1.6.0.0.

Thanks to jardleex for the patch.

Signed-off-by: Brian Foley <bpfoley@google.com>
Reviewed-by: Viktor Bachraty <vbachraty@google.com>