Use zlib compression level 3 in Haskell RPC code
This matches the behaviour of the python code and is faster but still
fairly efficient. For a large config.data, gzip -3 takes 25ms vs 40ms
for gzip -6 (ie ~60% of the runtime) and the compressed data is about
10% larger.

This is used during config.data replication to the master candidates
every time it's updated on the master.

Signed-off-by: Brian Foley <bpfoley@google.com>
Reviewed-by: Viktor Bachraty <vbachraty@google.com>