Tune getNodeInstances DRBD secondary computation
This is exercised by the luxi QueryInstances call when the sinst_cnt and
sint_list fields are used. This uses a lot of CPU and does a lot of
short-lived heap allocation on clusters with many instances.

The reimplementation allocates fewer temporary values, and does fewer
object lookups by UUID. The net effect is to reduce total heap allocation
from ~3.2GB to ~1.5GB, and CPU use from ~1200ms to ~770ms in a test
harness using a config with 1000 DRBD instances on 80 nodes.

Signed-off-by: Brian Foley <bpfoley@google.com>
Reviewed-by: Viktor Bachraty <vbachraty@google.com>