Slightly improve the JSScheduler/prop_jobFiltering test
For some reason, this test doesn't pass on Debian unstable. I don't know if it
passes on other versions of QuickCheck, but upon investigation, I think it was
always broken, but the use of stableCover hides it.

The problem is that the test wants min 4% coverage of the enumerated actions as
applyingFilter, among which is also Continue. But according to that function's
code and comments, Continue can never be the result of applyingFilter, which
means that this test passes randomly only then stableCover generates 2/10 True
cases (which can happen, since frequency AFAIK doesn't guarantee absolute
numbers); this makes the test pass as our required percent is very low (2/10 >
1.3/10 which is what we require). Removing the 'stableCover' call and using
instead 'cover' shows correctly that Cover is generated in 0% of cases (i.e.
never). Therefore, let's remove Cover from the list of required actions.

Additionally, slightly improve the test: instead of using '==>' and discarding
post-generation the unsuited tests, directly filter them during generation,
potentially making the test slightly faster.

Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>