Switch TagSet from type synonym to a newtype
This started from the fact that recent QuickCheck declares itself an arbitrary
instance "(Arbitrary a) => Arbitrary (Set a)", which conflicts with our
slightly more specific instance.  The easiest way to fix this is (especially
since this is test code) an OverlappingInstances language extension, but that
is deprecated in GHC 7.10 :(

The solution is then to change this to a proper type, instead of an alias.  In
the exercise, I found out that we were not really using TagSet, but instead
directly using Set, which is not nice. Plus many redefinitions at the call site
of the same arbitrary instance! Not happy with type synonims since they're just
smoke and mirrors…

Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>