Split getInstance into *ByUuid and *ByPartialName
This function (and getNode, next patch) were two pain points when I tried to
convert UUIDs to another type, since they're called via a single type
(currently String) that wants to dispatch to either UUID or name. In many cases
we know exactly what we want to search for, so having only this function is
overly generic. Not useful yet, but will be later for the string type
conversion.

Additionally change the name of the existing function getInstanceByName to
match the new functions better.

Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>