Ignore another bug in System.Time.diffClockTimes
Reading the source for that function is scary: it uses local time conversions
in the context of what should be absolute (not local) time, so of course it has
daylight savings time issues. It would be hard to work around those, so add a
FIXME and work around the test failure.

Found by using seed 7676442054323374465 on a 64-bit machine:

  timediffAdd: [Failed]
  *** Failed! Falsifiable (after 465 tests):
  Tue Jul 15 06:06:46 CEST 2008
  Wed Jan  5 15:00:19 CET 2028
  Fri Oct  5 18:35:48 CEST 2001
  Expected equality, but got mismatch
  expected: Fri Oct  5 18:35:48 CEST 2001
   but got: Fri Oct  5 19:35:48 CEST 2001

Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>