Fix KVM pinning tests to not depend on the local machine
Commit 8b2ec2f added unittests for KVM pinning, but it introduced a non-obvious
local dependency in the tests: the CPU_PINNING_OFF calls work by looking at the
(current) machine's core count, and pinning to all those CPUs. In order to make
this work independently from the test machine, we must also mock the result of
process.cpu_count(). Do this by using a core count that is very much unlikely
to ever be present in the real world.

Signed-off-by: Iustin Pop <iustin@google.com>
Signed-off-by: Viktor Bachraty <vbachraty@google.com>
Reviewed-by: Viktor Bachraty <vbachraty@google.com>