Prevent watcher from holding lock on verify group
The ganeti-watcher holds the group file lock for too long, until after
the execution of a group-verify-disk job. This locks for a long time if
there are other jobs already running and blocking the verify from
executing. When the lock is held, another ganeti-watcher run cannot be
scheduled, so this prevents the ganeti-watcher from running for several
minutes.

With this commit, the lock is released before running the VerifyDisks
operation, so even if the submitted job gets stuck in the Job Queue, a
subsequient ganeti-watcher run would still happen.

As an additional change, an incorrect docstring was also removed.

Signed-off-by: Federico Morg Pareschi <morg@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>