Improve missing job error handling in maintd
If luxi returned an error when getting statuses for any maintenance job
that was submitted while handling an incident and has not yet completed,
maintd would immediately fail, and could not recover from this condition
on subsequent rounds of maintenance.

This could happen if, for example, ganeti was manually shut down while
the job was running, and the job file was archived after ganeti
restarted.

Fix by treating missing/unparseable jobs as failed.

Signed-off-by: Brian Foley <bpfoley@google.com>
Reviewed-by: Viktor Bachraty <vbachraty@google.com>