Introduce a non-strict flag for group verify disk
The group verify disk ganeti job is executed every 5 minutes by a cron
job. This type of job tries to take all node locks in a nodegroup in
shared mode and can easily get stuck for a long time if there are
conflicting long-running jobs (like a create instance).

With this commit, we introduce a --no-strict flag to the watcher and
gnt_cluster verify-disks command to allow the group verify disk job to
only verify those nodes available for locking. The watcher will run a
fully-strict verify only every 30 minutes.

Signed-off-by: Federico Morg Pareschi <morg@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>