Use fork instead of spawnv in the watcher
With this commit, the ganeti-watcher now uses os.fork() instead of
spawning independent processes for each nodegroup to verify. With the
previous approach, there was very little memory being shared between
each subprocess and this introduced a significant and unnecessary
overhead. With this patch now the memory usage should be better
contained and should scale a bit better on larger clusters with multiple
nodegroups.

TRhe memory use on an x86_64 cluster with 4 node-groups on an x86_64 as
measured with ps_mem is:

        Private     Shared      Total                      (nprocs)
before: 92.6 MiB +   3.9 MiB =  96.5 MiB    ganeti-watcher (5)
after:  27.8 MiB +  15.7 MiB =  43.6 MiB    ganeti-watcher (5)

Signed-off-by: Federico Morg Pareschi <morg@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>