First implementation of the predictive scheduler
This is just a partial implementation of the static lock prediction system.
With this, Ganeti queue system tries to predict the locks of the running
jobs and schedule new jobs that are less likely to block.
It is a partial implementation which might cause starvation for jobs stuck
in the queue, following patches will address this issue.

Signed-off-by: Federico Morg Pareschi <morg@google.com>
Signed-off-by: Viktor Bachraty <vbachraty@google.com>
Reviewed-by: Viktor Bachraty <vbachraty@google.com>