Make FinalizeMigration{Src,Dst} more robust
Migration is supported on KVM/Xen hypervisors only. Finalize methods are
called in both success or suspected failure (the master node can't
certainly know about failures - e.g. in case of timeouts). Thus these
functions need to be as robust as possible in order to be able to
recover from failures - in case success=False the functions should be
conservative, trying to detect the actual state and idempotent. The
cleanup workaround works for XM, on XL it is a no-op.

Signed-off-by: Viktor Bachraty <vbachraty@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>