Make finalize_migration_{src,dst} a single op
FinalizeMigrationSource and FinalizeMigrationDst should compose an
single operation consisting of 2 idempotent steps, that can be always
both retried, similarly how they are used in AbortMigration() without
any reliance on the order of their execution. Otherwise we would have to
remember at which step did the migration fail in order to be able to
rollback/recover. Also state of the record should be updated only if
both succeeded, otherwise we won't know in case of recovery if the
record corresponds to the original state or the final state.

Signed-off-by: Viktor Bachraty <vbachraty@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>