StartInstance restores instance state if running
Add a side effect to StartInstance on Xen hypervisor that restores the
Xen configuration in case it is missing. This is a common failure
scenario during migrations. The generic backend checks if the instance
is running and queries the hypervisor about it's state. If it needs
recovery, the backend gathers the associated block devices and calls
RestoreInstance in the same manner as if it would call StartInstance in
case the instance wasn't already running.

Signed-off-by: Viktor Bachraty <vbachraty@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>