Fix for incorrect parsing of DRBD versions
Following issue #1194, this patch allows Ganeti to correctly
parse drbd versions that also include a dash in their k-fix
version component.

This technically makes Ganeti treat x.y.z.k the same as x.y.z-k
with regards to compatibility. For practicality we can assume that
these versions are the same, this is the intended behaviour as it is
unlikely that these two versions would be incompatible with each
other, if that ever were to happen.

Signed-off-by: Federico Morg Pareschi <morg@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>