Fix coexistence of location tags and non-DRBD instances
This addresses issue 1185, “hbal: IntMap.!: key -1 is not an element of
the map”. The issue is that the location tags code presumed all
instances have a primary and a secondary (i.e., they are DRBD).

The fix is to set the location score for non-DRBD instances to zero,
since this is what I understand the correct value: such an instance
cannot be optimized by moving, so its score is "perfect" (zero).

Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>