Improve the test data generation algorithms
This commit improves some of the test data generation in the ConfigData
dummy values of quickcheck. It generates more realistic instance, node
and group names by sampling them from the provided ConfigData. This
makes it possible to test for consistent data validation that might
depend on multiple names being present in the cluster configuration at
the same time.

Signed-off-by: Federico Morg Pareschi <morg@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>