Implement predictive queue cluster parameter
This commit adds the enabled_predictive_queue cluster parameter that
allows the cluster to specify if the predictive scheduler should be used
or not when ordering ganeti jobs in the queue. It also adds the correct
command line flags for the cluster init and cluster modify ganeti jobs.

As a default, if no option is specified upon cluster creation, the
predictive scheduler is automatically enabled.

Signed-off-by: Federico Morg Pareschi <morg@google.com>
Reviewed-by: Viktor Bachraty <vbachraty@google.com>