Prohibit disk removal w/o hotplug on live instance
Prior to this commit, removing a disk from a live instance without
using hotplug succeeded incorrectly leaving stale disks behind. This is
due to the fact that the hypervisor holds the device open while Ganeti
tries to shut it down. The operation fails but Ganeti continues anyway,
removing the disk from the config and leaving the disk stale.

This patch adds an appropriate check before removing a disk from a live
instance.

Signed-off-by: Yiannis Tsiouris <tsiou@grnet.gr>
Signed-off-by: Filippos Giannakos <philipgian@grnet.gr>
Reviewed-by: Viktor Bachraty <vbachraty@google.com>