Add optStaticKvmNodeMemory to HTools and IAllocator
This is an attempt to fix balancing on 2.16 using KVM hypervisor. The problem
with KVM is, that it reports 'active' memory from /proc/memstat for node
memory a.k.a. nMem a.k.a. memory_dom0.

This is plain wrong, as 'active' memory in Linux includes memory from all
running processes, including instances (qemu processes). We can't calculate
the node memory on the hypervisor side, simply subtracting RES used by qemu
processes would give us only approximate value and still it would be the actively
used memory rather than a upper hard limit as in dom0 size in Xen. As KVM
does not provide proper isolation for node memory, the only thing we can do is
to instruct htools to reserve a static value for node memory.

This is solved in 2.17 in patch f971341471 implementing the obtainNodeMemory
function that takes the live dom0 size in Xen, but uses 'hv_state' for exporting
the node size configurable parameter for KVM. I've opted for _not_ backporting
the cluster parameter from 2.17 for these reasons:

* this way the change is isolated to hbal, no need to change cluster config
* the obtainNodeMemory is for Luxi/Rapi only, this flag works with any backend
* the goal is to fix balancing as it is, without adding support for memory
overcommitment, doing so would require backporting a bunch of other dependencies:
d2bfc50608, f971341471 and possibly more.

The flag is ignored on Xen, and on KVM is ste by default to 4G, so the change
should be backward compatible (no need to specify this flag unless you need
a different value).

Signed-off-by: Viktor Bachraty <vbachraty@google.com>