Redesign Ganeti memory model
The existing memory model in htools was reporting incorrect node memory,
missing memory and also available (free) memory on KVM that resulted in some
cases in overcramming nodes (overcommiting memory without having swap to evict
pages from RAM) leading to OOM kills.

This model is equivalent to the model proposed in issue 127, but doesn't
introduce overcommitment (swap or balooning) support nor does it depend
on Ganeti 2.17 features.

The goal is to distinguish between reported (state-of-world) values and recorded
(state-of-record) values and not to mix them arbitrarily. To estimate free
memory we always need to take the minimum of measured/calculated values using
those two states (root causes of the differences explained in missing memory) in
order to be on the safe side.

Signed-off-by: Viktor Bachraty <vbachraty@google.com>