Force "sudo: false" on Travis-CI config file.
We should use the container build to properly isolate our dependencies. In order
to do this we have to run the build using "sudo: false". The current
configuration file works fine for recent forks because the default for them is 
"sudo: false", while the default for older repos is "sudo: required" (see [1]).

[1]: https://docs.travis-ci.com/user/ci-environment/#Virtualization-environments 

Signed-off-by: Rafael Marinheiro <marinheiro@google.com>