Force CleanupInstance always on InstanceShutdown
This commit fixes the bug described in #52 by always trying to cleanup
an instance's runtime files and block devices on InstanceShutdown. The
buggy scenario is:
- A user issues a 'poweroff' from inside the VM
- The admin or a RAPI client does a 'gnt-instance shutdown' (or even
worse a 'gnt-instance remove' which, in turn, issues an instance
shutdown) *before* the watcher notices the 'ERROR_down' state of the
VM
- The instance status changes to 'ADMIN_down' (or gets removed) but the
cleanup is never performed. Thus, the CTRL_DIR, CONF_DIR, NIC_DIR,
instance_disks, etc. are left stale.

This can be reproduced on clusters that have the 'user_shutdown' option
disabled.

Signed-off-by: Yiannis Tsiouris <tsiou@grnet.gr>
Signed-off-by: Alexandros Kiousis <alexk@noc.grnet.gr>