Fix documentation builds with Sphinx >= 1.4
We are knowingly overriding sphinx's build-in manpage role, so we need
to suppress sphinx's warnings; otherwise, when run with the `-W' switch,
sphinx will error out.

We also need to replace the unicode ellipses in doc/design-query2.rst
with ascii periods, as sphinx amits another warning for not being able
to properly highlight these lines as Python code.

Signed-off-by: Apollon Oikonomopoulos <apoikos@debian.org>