Update TestHelper conInfo parser to be exhaustive
This change is necessary if we want to compile Ganeti tests with
-Werror. Two data constructors have been added to support list of names:
GadtC and RecGadtC. By default, we identify the name as the first
element in the list and ignore the rest. For any other corner case, we
simply turn the result type into a Maybe type and in case of error we
let the compiler fail by accessing a Nothing with a fromJust.

This is appropriate because it's just a Test and not production code, we
do not care if we're crashing the test generator in case of error.

Signed-off-by: Federico 'Morg' Pareschi <morg@google.com>