utils.livelock: use portable struct flock type
From fcntl(2):
     struct flock {
         ...
         short l_type;    /* Type of lock: F_RDLCK,
                             F_WRLCK, F_UNLCK */
         short l_whence;  /* How to interpret l_start:
                             SEEK_SET, SEEK_CUR, SEEK_END */
         off_t l_start;   /* Starting offset for lock */
         off_t l_len;     /* Number of bytes to lock */
         pid_t l_pid;     /* PID of process blocking our lock
                             (set by F_GETLK and F_OFD_GETLK) */
         ...
     };

On 64-bit systems, off_t is always 64 bits long ("long"/"long long"). On
32-bit systems however, depending on whether large file support is
enabled or not, it may be 64 bits ("long long") or 32 bits ("long")
long.

The code in LiveLock.__init__ would always assume off_t to be "long",
breaking on 32-bit systems with LFS support. Fix this by picking the
correct type to use depending on the existence of os.O_LARGEFILE.

Note that LFS is enabled almost universally these days and it would be
safe to just use "long long" unconditionally, but it doesn't harm to
make the actual check.

Signed-off-by: Apollon Oikonomopoulos <apoikos@debian.org>