Fix Haskell daemon '-b' help text
This refers to, and fixes #974.

Looking at the source code, the previous "default depends on cluster
configuration" refers to the address _family_, not the address itself
(which is always 'any' as in INADDR_ANY or INADDR6_ANY). Let's correct
the help text and also the docstring for that parameter (which also
implied an override of a specific address).

Signed-off-by: Iustin Pop <iustin@k1024.org>