Cabal 2.0 & 2.2 compatibility
Cabal 2.0 (included in GHC 8.2) made various internal changes mandating
components to be passed in various places, Macros.generate being one of
them. In our case, it is okay to use the main library component for
dependency generation.

Additionally, in Cabal 2.2 Distribution.PackageDescription.Parse was
removed, so we switch to using .Parsec instead.