template-haskell 2.11 & 2.12 compatibility
Template Haskell 2.11.0.0 changed the DataD family of constructors to
accept a new field of type `Maybe Kind`, which we can safely set to
`Nothing`.

`Strict` has been replaced with `Bang`, so we re-implement it as
`myNotStrict NoSourceUnpackedness NoSourceStrictness`.

Additionally, the VarI constructor no longer takes a `Fixity` field.

Finally, Template Haskell 2.12.0.0 requires names to be passed to DataD
as DerivClauses, so we add a little shim function, `derivesFromNames` to
do the right thing depending on TH version. Ship this function in a
separate Ganeti.THH.Compat module to avoid enabling CPP for THH.hs.

And since we're supporting GHC 8.0+ anyway, we can bump
template-haskell's minimum version to 2.11.0.0.

This fixes the Template Haskell part for #1287 and #1312.

Signed-off-by: Apollon Oikonomopoulos <apoikos@debian.org>