Make all instances of Monoid, instances of Semigroup
This is a compatibility fix for base-4.11/GHC 8.4. In base-4.11,
Semigroup is a superclass of Monoid and Semigroup itself is exported in
Prelude.

Make all our Monoid instances also instances of Semigroup and define
mappend in terms of (<>).

Note that to keep compatibility with base-4.9 and avoid duplicate import
warnings, we do a qualified import of Data.Semigroup as Sem. This can be
removed once we drop compatibility for anything earlier than base-4.11.