Add hpack package.yaml
Replicate most of our cabal's logic. hpack autogenerate's the list of
exposed modules and cabal will now call ghc with the proper package-ids.
The latter in addition to the fact that GHC >= 8.0 auto-generates the
MIN_VERSION_* macros from the given package-ids means that we can get
completely rid of the custom cabal logic in cabal/.