testutils: allow skipping IPv6 tests when not supported
Add the RequiresIPv6 test decorator that attempts to check whether IPv6
is supported on the host system, and will skip tests requiring IPv6
support in case it's not.

This is required for the testsuite to run on potentially restricted
environments, for instance on Travis docker containers.

Signed-off-by: Apollon Oikonomopoulos <apoikos@debian.org>
(cherry picked from commit 14e34da8847e4410fd3db0f850b0c6a265324723)
Signed-off-by: Iustin Pop <iustin@k1024.org>