daemon-util: also match processes by name when using s-s-d
As of dpkg 1.19.4, start-stop-daemon will refuse to match a running
process by pidfile only, when the pidfile is not owned by root. It does
so on security grounds, as a user-controlled pidfile could be used to
kill any process on the system.

When running with user separation enabled, most Ganeti daemons write
their PID files as regular users, which means that process control with
s-s-d is currently broken. Fix this by also matching processes by name,
in addition to the PID file.

Fixes #1322.

Signed-off-by: Apollon Oikonomopoulos <apoikos@debian.org>