Fix openvswitch port setup when no untagged VLANs are used
OpenVSwitch port setup fails when a port is only configured to carry
tagged VLANs, due to the VLAN's configuration string being malformed.
This patch fixes these cases. Here's an example that triggers this
without this patch, where a configured NIC is set up in openvswitch
mode with tagged frames only:

  0:link=vswitch0,mode=openvswitch,vlan=:120:1200

Signed-off-by: George Diamantopoulos <gdiamantopoulos@modulus.gr>