KVM: fix VNC TLS handling for QEMU 3.1
Since QEMU 2.5, VNC TLS handling has been refactored using the `-object
tls-creds-*` parameters. QEMU 3.1 completely removed the support for
`-vnc tls`, so we need to switch to the new syntax.

Note that this places a lower bound on QEMU 2.5 for those wishing to use
VNC + TLS. 2.5.0 was released more than 3 years ago, so people should be
using it already and it doesn't make sense to try to be backwards
compatible here.

This is part of #1338.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>