ganeti-kvm-poweroff: add Default-Start runlevels
The ganeti-kvm-poweroff script is lacking Default-Start runlevels. While
this is okay with sysvinit, it doesn't work with systemd because the
latter keeps state of whether a service was started before stopping it.

Fix this by populating Default-Start on the LSB header. Note that start
is noop already, so this is harmless.

Thanks to @pboguslawski.

Fixes #1288.