Restore compatibility with TH < 2.11
stable-2.16 needs to remain backwards compatible with GHC 7.6. This
commit restores compatibility with TH < 2.11 by factoring out DataD and
InstanceD construction to a set of functions and by abstracting DataD
pattern matching to extract constructor information.

We also move the defition of myNotStrict to Ganeti.THH.Compat and make
it conditional to the template-haskell version where Bang appeared.

Finally, we add some CPP sugar in Ganeti.THH.Types to account for the
arity change in VarI. Unfortunately this is not easy to factor out to
Ganeti.THH.Compat without moving the whole function, which would be
uglier.

We can now restore the lower bound for template-haskell to 2.7.

Signed-off-by: Apollon Oikonomopoulos <apoikos@dmesg.gr>