Convert TagSet to a "proper" type
Type aliases have multiple problems, one of which hits us in the form
of (upstream/default) Arbitrary instances. For TagSet, that means the
arbitrary instance will be a set of random strings, which doesn't give
as good a coverage as our desired tag formats.

Changing this to its own proper type is much simpler than I thought,
since the JSON instance is the heavy-usage path; there is only one
other case where we need to adjust to the new type.

Tests fail in the same way as before this patch ('jobFiltering'
property fails coverage).

Signed-off-by: Iustin Pop <iustin@k1024.org>