Remove `==>` from JQScheduler tests
Personal gripe: `==>` should only be used as last resort. It's also
one of the reasons people introduce coverage checks, because `==>`
skews your actual tests as well.

Whenever possible, `suchThat` should be used, so that the generated
values are appropriate from the start, rather then generating random
values and then skipping the test.

In this particular module, the use of `==>` was for really arbitrary
reasons, where all uses could be directly tested in the generation
step…

Signed-off-by: Iustin Pop <iustin@k1024.org>